% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_balance_data.R
\name{clean_balance_data}
\alias{clean_balance_data}
\title{After extracting balance data, clean it by removing redundant data}
\usage{
clean_balance_data(data, remove_words = "Totaal|Balans|Kas")
}
\arguments{
\item{data}{data.frame}

\item{remove_words}{character}
}
\value{

}
\description{
After extracting balance data, clean it by removing redundant data
}
