% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_balance_data.R
\name{extract_balance_data}
\alias{extract_balance_data}
\title{Extract Balance sheet data from FD website.}
\usage{
extract_balance_data(website)
}
\arguments{
\item{website}{character}
}
\value{
data.frame
}
\description{
Extract Balance sheet data from FD website.
}
