% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_fd_data.R
\name{scrape_fd_data}
\alias{scrape_fd_data}
\title{Scrape Balance Sheet data from multiple FD websites}
\usage{
scrape_fd_data(websites)
}
\arguments{
\item{websites}{Character. Vector of urls.}
}
\value{

}
\description{
Scrape Balance Sheet data from multiple FD websites
}
