% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_fd_url.R
\name{scrape_fd_url}
\alias{scrape_fd_url}
\title{Scrape a url from the FD website.}
\usage{
scrape_fd_url(url)
}
\arguments{
\item{url}{character}
}
\value{

}
\description{
Scrape a url from the FD website.
}
