% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_fd_urls.R
\name{scrape_fd_urls}
\alias{scrape_fd_urls}
\title{Wrapper around FD url to scrape multiple websites at once}
\usage{
scrape_fd_urls(urls)
}
\arguments{
\item{urls}{Character. Link to website.}
}
\value{

}
\description{
Wrapper around FD url to scrape multiple websites at once
}
