% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_balance_data.R
\name{extract_balance_data}
\alias{extract_balance_data}
\title{Extract Balance sheet data from the FD website.}
\usage{
extract_balance_data(website, remove_chars = c("", "--", "0"), dec = ",")
}
\arguments{
\item{website}{character. url.}

\item{remove_chars}{character. What characters do you want to
remove from the balance sheet data. Default: c("", "--", "0").}

\item{dec}{character. Default: ','}
}
\value{
data.frame
}
\description{
Extract Balance sheet data from the FD website.
}
\examples{
data <- extract_balance_data("https://beurs.fd.nl/noteringen/11755/ahold-delhaize/kerncijfers")
print(data)
}
