% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_fd_urls.R
\name{scrape_fd_urls}
\alias{scrape_fd_urls}
\title{Wrapper around \code{scrape_fd_url} to find balance sheet data from
specific companies.}
\usage{
scrape_fd_urls(urls)
}
\arguments{
\item{urls}{Character. Link to website.}
}
\value{

}
\description{
Wrapper around \code{scrape_fd_url} to find balance sheet data from
specific companies.
}
\examples{
# urls to scrape weblinks from
urls <- c(
"https://beurs.fd.nl/analyse/amsterdam/aex/",
"https://beurs.fd.nl/analyse/amsterdam/midkap/"
)
websites <- scrape_fd_urls(urls)
print(websites)
}
